dbt-dag/dags/dbt_dag.py
from datetime import datetime

from cosmos import DbtDag, ProjectConfig, ProfileConfig
from cosmos.profiles import SnowflakeUserPasswordProfileMapping

profile_config = ProfileConfig(
    profile_name = "default",
    target_name = "dev",
    profile_mapping = SnowflakeUserPasswordProfileMapping(
        conn_id = "snowflake_conn",
        profile_args = {"database": "dbt_db", "schema": "dbt_schema"},
    )

)

dbt_snowflake_dag = DbtDag(
    project_config = ProjectConfig("/usr/local/airflow/dags/data_pipeline"),
    operator_args = {"install_deps": True},
    profile_config = profile_config,
    schedule = "@daily",
    start_date = datetime(2023, 1, 1),
    catchup = False,
    dag_id = "dbt_dag",
)